+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 - yellowstone-grpc-client 13.3.0
 - yellowstone-grpc-client-simple-12.3.1
+- yellowstone-grpc-client-nodejs-5.1.0
 
 ### Features